Update details of Xen support
There are two nova drivers:
* XenAPI driver for XenServer
* Libvirt + Xen

Currently only the XenAPI is a support Nova driver. The initial
prototype was available in the Austin release, but started really
working in the Bexar release.

The XenServer ci is testing the XenAPI driver, not the Xen+libvirt
driver.

None of these are whole Citrix contributions, the XenAPI is mostly
maintained by Rackspace, but also a few others. The libvirt based driver
for Xen has never had any support from Citrix, and has recently had the
most support from SUSE.