Fix broken links to driver docs
The documentation team recently killed the 'trunk' guide links which
were used by a lot of drivers in driverlog.

This patch goes through and updates any config-reference link that
was pointing at 'trunk' to the link of the latest supported version.

There are about 4 links for which the link was broken prior to this
change, and these have been left as-is so they are easier to find and fix.

See also related patch: https://review.openstack.org/#/c/179987/ which
introduces an HTTP redirect to fix old links to trunk.

Change-Id: I9974580590c4a292b36b73ecc91bed421cabcfae
Closes-Bug: 1451501