Update some nova driver contacts
This is just updating some of the nova driver contacts to
replace people no longer working on these drivers. For
ironic and the kvm/qemu/lxc ones, those don't have specific
owners, they are a team so I've adjusted accordingly.

There is a questionable gap in CI coverage (we don't actually
test with KVM nor LXC), and release support (not sure how that
is determined or updated), but I don't plan on dealing with those
here.

Change-Id: I3dc384d4f78a4afe612e7f8e74449e7d1b1c6fbc