Add Queens, Update VMware Driver Info
Several of the VMware drivers have outdated information
(mainly releases and maintainers).  This patch updates them
with more current information.  Among the releases these drivers
are included in is Queens, which was released this week but is
not yet listed in default_data.json.  This patch adds it as well
so we don't have to patch the driver listings a second time.

Change-Id: Id5482c27e371f67b782772d03c74e926a2804795